refactor(api): move initiatives to modules/ and share the page query and envelope
