+++ apps/api/src/issues/store.ts
@@ -50,7 +50,7 @@ import { notifyIssueChange } from '../notifications/store';
 import { emitWebhookEvent } from '../webhooks/emit';
 import { getAssignTriggerAgent, isProjectAgent } from '#modules/agents/core/service';
 import { deleteThreadsWhere } from '#modules/agents/core/runtime/memory';
-import { getInitiativeProjectId } from '../initiatives/store';
+import { getInitiativeProjectId } from '#modules/initiatives/service';
 import { getCycleRef } from '#modules/cycles/service';
 import { getMembership } from '../members/store';
 import { enqueueAgentRun } from '#modules/agents/core/run-queue';
